server: add projectHelper utility class
