+++ language-server/src/features/commands/parsedTypeInfoRequest.ts
@@ -1,31 +1,30 @@
 import { Connection } from 'vscode-languageserver'
 import { Logger } from 'winston'
 import { ParsedTypeInfoRequest, ParsedTypeInfoRequestResponse } from '../../events'
-import { LoadFolder } from '../../mod/loadfolders'
 import { ProjectManager } from '../../projectManager'
 import { Provider } from '../provider'
 import * as winston from 'winston'
 import * as tsyringe from 'tsyringe'
 import { LogToken } from '../../log'
+import { ProjectHelper } from '../utils/project'
 
 @tsyringe.injectable()
-export class ParsedTypeInfoRequestHandler extends Provider {
+export class ParsedTypeInfoRequestHandler implements Provider {
   private logFormat = winston.format.printf(
     (info) => `[${info.level}] [${ParsedTypeInfoRequestHandler.name}] ${info.message}`
   )
   private readonly log: winston.Logger
 
   constructor(
-    loadFolder: LoadFolder,
-    projectManager: ProjectManager,
+    private readonly projectManager: ProjectManager,
+    private readonly projectHelper: ProjectHelper,
     @tsyringe.inject(LogToken) baseLogger: winston.Logger
   ) {
-    super(loadFolder, projectManager)
     this.log = winston.createLogger({ transports: baseLogger.transports, format: this.logFormat })
   }
 
   listen(connection: Connection): void {
-    connection.onRequest(ParsedTypeInfoRequest, this.wrapExceptionStackTraces(this.onRequest.bind(this)))
+    connection.onRequest(ParsedTypeInfoRequest, this.projectHelper.wrapExceptionStackTraces(this.onRequest.bind(this)))
   }
 
   protected getLogger(): Logger {
